requirements.txt
alembic==1.6.2
auth0-python==3.16.2
Authlib==0.15.4
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
cryptography==3.4.7
ecdsa==0.14.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-jose==3.2.0
requests==2.25.1
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.15
typing==3.7.4.3
urllib3==1.26.4
Werkzeug==2.0.0
WTForms==2.3.3
